LPS-191331 Move methods to base class
